+++ Prowl.Runtime/Math/Plane.cs
@@ -30,7 +30,6 @@ SOFTWARE.
 
 using System;
 using System.Diagnostics;
-using System.Numerics;
 using System.Runtime.Serialization;
 
 namespace Prowl.Runtime
@@ -43,7 +42,7 @@ namespace Prowl.Runtime
         /// <param name="point">The point to check with</param>
         /// <param name="plane">The plane to check against</param>
         /// <returns>Greater than zero if on the positive side, less than zero if on the negative size, 0 otherwise</returns>
-        public static float ClassifyPoint(ref Vector3 point, ref Plane plane)
+        public static double ClassifyPoint(ref Vector3 point, ref Plane plane)
         {
             return point.X * plane.Normal.X + point.Y * plane.Normal.Y + point.Z * plane.Normal.Z + plane.D;
         }
@@ -54,10 +53,10 @@ namespace Prowl.Runtime
         /// <param name="point">The point to check</param>
         /// <param name="plane">The place to check</param>
         /// <returns>The perpendicular distance from the point to the plane</returns>
-        public static float PerpendicularDistance(ref Vector3 point, ref Plane plane)
+        public static double PerpendicularDistance(ref Vector3 point, ref Plane plane)
         {
             // dist = (ax + by + cz + d) / sqrt(a*a + b*b + c*c)
-            return (float)Math.Abs((plane.Normal.X * point.X + plane.Normal.Y * point.Y + plane.Normal.Z * point.Z)
+            return (double)Math.Abs((plane.Normal.X * point.X + plane.Normal.Y * point.Y + plane.Normal.Z * point.Z)
                                     / Math.Sqrt(plane.Normal.X * plane.Normal.X + plane.Normal.Y * plane.Normal.Y + plane.Normal.Z * plane.Normal.Z));
         }
     }
@@ -73,7 +72,7 @@ namespace Prowl.Runtime
     {
         #region Public Fields
 
-        public float D;
+        public double D;
 
         public Vector3 Normal;
 
@@ -88,7 +87,7 @@ namespace Prowl.Runtime
 
         }
 
-        public Plane(Vector3 normal, float d)
+        public Plane(Vector3 normal, double d)
         {
             Normal = normal;
             D = d;
@@ -104,7 +103,7 @@ namespace Prowl.Runtime
             D = -(Vector3.Dot(Normal, a));
         }
 
-        public Plane(float a, float b, float c, float d)
+        public Plane(double a, double b, double c, double d)
             : this(new Vector3(a, b, c), d)
         {
 
@@ -115,43 +114,43 @@ namespace Prowl.Runtime
 
         #region Public Methods
 
-        public float Dot(Vector4 value)
+        public double Dot(Vector4 value)
         {
             return ((((this.Normal.X * value.X) + (this.Normal.Y * value.Y)) + (this.Normal.Z * value.Z)) + (this.D * value.W));
         }
 
-        public void Dot(ref Vector4 value, out float result)
+        public void Dot(ref Vector4 value, out double result)
         {
             result = (((this.Normal.X * value.X) + (this.Normal.Y * value.Y)) + (this.Normal.Z * value.Z)) + (this.D * value.W);
         }
 
-        public float DotCoordinate(Vector3 value)
+        public double DotCoordinate(Vector3 value)
         {
             return ((((this.Normal.X * value.X) + (this.Normal.Y * value.Y)) + (this.Normal.Z * value.Z)) + this.D);
         }
 
-        public void DotCoordinate(ref Vector3 value, out float result)
+        public void DotCoordinate(ref Vector3 value, out double result)
         {
             result = (((this.Normal.X * value.X) + (this.Normal.Y * value.Y)) + (this.Normal.Z * value.Z)) + this.D;
         }
 
-        public float DotNormal(Vector3 value)
+        public double DotNormal(Vector3 value)
         {
             return (((this.Normal.X * value.X) + (this.Normal.Y * value.Y)) + (this.Normal.Z * value.Z));
         }
 
-        public void DotNormal(ref Vector3 value, out float result)
+        public void DotNormal(ref Vector3 value, out double result)
         {
             result = ((this.Normal.X * value.X) + (this.Normal.Y * value.Y)) + (this.Normal.Z * value.Z);
         }
 
         public void Normalize()
         {
-            float factor;
+            double factor;
             Vector3 normal = Normal;
             Normal = Vector3.Normalize(Normal);
-            factor = (float)Math.Sqrt(Normal.X * Normal.X + Normal.Y * Normal.Y + Normal.Z * Normal.Z) /
-                    (float)Math.Sqrt(normal.X * normal.X + normal.Y * normal.Y + normal.Z * normal.Z);
+            factor = (double)Math.Sqrt(Normal.X * Normal.X + Normal.Y * Normal.Y + Normal.Z * Normal.Z) /
+                    (double)Math.Sqrt(normal.X * normal.X + normal.Y * normal.Y + normal.Z * normal.Z);
             D = D * factor;
         }
 
@@ -164,10 +163,10 @@ namespace Prowl.Runtime
 
         public static void Normalize(ref Plane value, out Plane result)
         {
-            float factor;
+            double factor;
             result.Normal = Vector3.Normalize(value.Normal);
-            factor = (float)Math.Sqrt(result.Normal.X * result.Normal.X + result.Normal.Y * result.Normal.Y + result.Normal.Z * result.Normal.Z) /
-                    (float)Math.Sqrt(value.Normal.X * value.Normal.X + value.Normal.Y * value.Normal.Y + value.Normal.Z * value.Normal.Z);
+            factor = (double)Math.Sqrt(result.Normal.X * result.Normal.X + result.Normal.Y * result.Normal.Y + result.Normal.Z * result.Normal.Z) /
+                    (double)Math.Sqrt(value.Normal.X * value.Normal.X + value.Normal.Y * value.Normal.Y + value.Normal.Z * value.Normal.Z);
             result.D = value.D * factor;
         }
 
@@ -208,7 +207,7 @@ namespace Prowl.Runtime
 
         internal PlaneIntersectionType Intersects(ref Vector3 point)
         {
-            float distance;
+            double distance;
             DotCoordinate(ref point, out distance);
 
             if (distance > 0)
